removed feeds from index.tsx
